reorganization and a bunch of documentation
