+++ src/Swarm/Game/World/Typecheck.hs
@@ -36,6 +36,35 @@ import Swarm.Util.Erasable
 import Prelude hiding (lookup)
 
 ------------------------------------------------------------
+-- Type classes for monoidal world values
+
+-- We could use Semigroup and Monoid, but we want to use the two
+-- classes separately and make instances for base types, so it's
+-- cleaner to just make our own classes (the instances would be
+-- orphans if we used Semigroup and Monoid).
+
+class Empty e where
+  empty :: e
+
+instance Empty CellVal where
+  empty = CellVal mempty mempty mempty
+
+class Over m where
+  (<+>) :: m -> m -> m
+
+instance Over Bool where
+  _ <+> x = x
+
+instance Over Integer where
+  _ <+> x = x
+
+instance Over Double where
+  _ <+> x = x
+
+instance Over CellVal where
+  CellVal t1 e1 r1 <+> CellVal t2 e2 r2 = CellVal (t1 <> t2) (e1 <> e2) (r1 <> r2)
+
+------------------------------------------------------------
 -- Type class for type-indexed application
 
 infixl 1 $$
@@ -43,13 +72,19 @@ class Applicable t where
   ($$) :: t (a -> b) -> t a -> t b
 
 ------------------------------------------------------------
--- Type-indexed constants
+-- Distinguishing functions and non-functions at the type level
+
+-- In several places, for efficiency we will require something to be
+-- not a function, which we can enforce using the 'NotFun' constraint.
 
 type family IsFun a where
   IsFun (_ -> _) = 'True
   IsFun _ = 'False
 
 type NotFun a = IsFun a ~ 'False
+
+------------------------------------------------------------
+-- Type-indexed constants
 
 -- Includes language built-ins as well as combinators we will use
 -- later as a compilation target.
@@ -86,12 +121,16 @@ data Const :: Type -> Type where
   CRot :: Rot -> Const (World a -> World a)
   COver :: (Over a, NotFun a) => Const (a -> a -> a)
   CEmpty :: (Empty a, NotFun a) => Const a
+  -- Combinators generated during elaboration
   K :: Const (a -> b -> a)
   S :: Const ((a -> b -> c) -> (a -> b) -> a -> c)
   I :: Const (a -> a)
   B :: Const ((b -> c) -> (a -> b) -> a -> c)
   C :: Const ((a -> b -> c) -> b -> a -> c)
-  Φ :: Const ((a -> b -> c) -> (d -> a) -> (d -> b) -> (d -> c)) -- Phoenix, aka liftA2
+  -- Phoenix combinator, aka liftA2.  Including this combinator in the
+  -- target set is not typical, but it turns out to be very helpful in
+  -- elaborating the "over" operation.
+  Φ :: Const ((a -> b -> c) -> (d -> a) -> (d -> b) -> (d -> c))
 
 deriving instance Show (Const ty)
 
@@ -113,17 +152,21 @@ c1 .$$. c2 = embed c1 $$ embed c2
 ------------------------------------------------------------
 -- Intrinsically typed core language
 
+-- | Type-level list append.
 type family Append (xs :: [k]) (ys :: [k]) :: [k] where
   Append '[] ys = ys
   Append (x ': xs) ys = x ': Append xs ys
 
--- Typed de Bruijn indices.
+-- | Type- and context-indexed de Bruijn indices. (v :: Idx g a) means
+--   v represents a variable with type a in a type context g.
 data Idx :: [Type] -> Type -> Type where
   VZ :: Idx (ty ': g) ty
   VS :: Idx g ty -> Idx (x ': g) ty
 
 deriving instance Show (Idx g ty)
 
+-- | A variable valid in one context is also valid in another extended
+--   context with additional variables.
 weakenVar :: forall h g a. Idx g a -> Idx (Append g h) a
 weakenVar VZ = VZ
 weakenVar (VS x) = VS (weakenVar @h x)
@@ -145,6 +188,8 @@ instance Applicable (TTerm g) where
 instance HasConst (TTerm g) where
   embed = TConst
 
+-- | A term valid in one context is also valid in another extended
+--   context with additional variables (which the term does not use).
 weaken :: forall h g a. TTerm g a -> TTerm (Append g h) a
 weaken (TVar x) = TVar (weakenVar @h x)
 weaken (TLam t) = TLam (weaken @h t)
@@ -154,6 +199,7 @@ weaken (TConst c) = TConst c
 ------------------------------------------------------------
 -- Errors
 
+-- | Errors that can occur during typechecking/elaboration.
 data CheckErr where
   UnknownErr :: Int -> CheckErr
   ApplyErr :: Some (TTerm g) -> Some (TTerm g) -> CheckErr
@@ -163,6 +209,7 @@ deriving instance Show CheckErr
 ------------------------------------------------------------
 -- Type representations
 
+-- | Base types.
 data Base :: Type -> Type where
   BInt :: Base Integer
   BFloat :: Base Double
@@ -171,6 +218,8 @@ data Base :: Type -> Type where
 
 deriving instance Show (Base ty)
 
+-- | Testing base type representations for equality to yield reflected
+--   type-level equalities.
 instance TestEquality Base where
   testEquality BInt BInt = Just Refl
   testEquality BFloat BFloat = Just Refl
@@ -178,6 +227,8 @@ instance TestEquality Base where
   testEquality BCell BCell = Just Refl
   testEquality _ _ = Nothing
 
+-- | Type representations indexed by the corresponding host language
+--   type.
 data TTy :: Type -> Type where
   TTyBase :: Base t -> TTy t
   (:->:) :: TTy a -> TTy b -> TTy (a -> b)
@@ -199,6 +250,8 @@ pattern TTyCell = TTyBase BCell
 
 deriving instance Show (TTy ty)
 
+-- | Testing type representations for equality to yield reflected
+--   type-level equalities.
 instance TestEquality TTy where
   testEquality (TTyBase b1) (TTyBase b2) = testEquality b1 b2
   testEquality (TTyWorld b1) (TTyWorld b2) =
@@ -207,6 +260,9 @@ instance TestEquality TTy where
       Nothing -> Nothing
   testEquality _ _ = Nothing
 
+-- | Check that a particular type has an 'Eq' instance, and run a
+--   computation in a context provided with an 'Eq' constraint. The
+--   other @checkX@ functions are similar.
 checkEq :: (Has (Throw CheckErr) sig m) => TTy ty -> ((Eq ty, NotFun ty) => m a) -> m a
 checkEq (TTyBase BBool) a = a
 checkEq (TTyBase BInt) a = a
@@ -242,6 +298,9 @@ checkOver _ _ = throwError (UnknownErr 5)
 ------------------------------------------------------------
 -- Existential wrappers
 
+-- | Wrap up a type-indexed thing to hide the type index, but package
+--   it with a 'TTy' which we can pattern-match on to recover the type
+--   later.
 data Some :: (Type -> Type) -> Type where
   Some :: TTy α -> t α -> Some t
 
@@ -256,21 +315,30 @@ pattern SomeTy :: TTy α -> SomeTy
 pattern SomeTy α = Some α (F.Const ())
 {-# COMPLETE SomeTy #-}
 
+------------------------------------------------------------
+-- Type inference/checking + elaboration
+
+type WExpMap = Map Text (Some (TTerm '[]))
+
+-- | Type contexts, indexed by a type-level list of types of all the
+--   variables in the context.
 data Ctx :: [Type] -> Type where
   CNil :: Ctx '[]
   CCons :: Text -> TTy ty -> Ctx g -> Ctx (ty ': g)
 
+-- | Look up a variable name in the context, returning a type-indexed
+--   de Bruijn index.
 lookup :: (Has (Throw CheckErr) sig m) => Text -> Ctx g -> m (Some (Idx g))
 lookup _ CNil = throwError (UnknownErr 6)
 lookup x (CCons y ty ctx)
   | x == y = return $ Some ty VZ
   | otherwise = mapSome VS <$> lookup x ctx
 
-------------------------------------------------------------
--- Type inference/checking + elaboration
-
-type WExpMap = Map Text (Some (TTerm '[]))
-
+-- | Check that a term has a given type, and if so, return a
+--   corresponding elaborated and type-indexed term.  Note that this
+--   also deals with subtyping: for example, if we check that the term
+--   @3@ has type @World Int@, we will get back a suitably lifted
+--   value (/i.e./ @const 3@).
 check ::
   ( Has (Throw CheckErr) sig m
   , Has (Reader EntityMap) sig m
@@ -287,21 +355,35 @@ check e ty t = do
     Nothing -> throwError (UnknownErr 7)
     Just Refl -> return t'
 
--- | XXX
+-- | Get the underlying base type of a term which either has a base
+--   type or a World type.
 getBaseType :: Some (TTerm g) -> SomeTy
 getBaseType (Some (TTyWorld ty) _) = SomeTy ty
 getBaseType (Some ty _) = SomeTy ty
 
--- | XXX Application is where we deal with lifting + promotion.
+-- | Apply one term to another term, automatically handling promotion
+--   and lifting, via the fact that World is Applicative.  That is,
+--   (1) if a term of type T is used where a term of type World T is
+--   expected, it will automatically be promoted (by an application of
+--   const); (2) if a function of type (T1 -> T2 -> ... -> Tn) is
+--   applied to any arguments of type (World Ti), the function will be
+--   lifted to (World T1 -> World T2 -> ... -> World Tn).
 apply :: (Has (Throw CheckErr) sig m) => Some (TTerm g) -> Some (TTerm g) -> m (Some (TTerm g))
+-- Normal function application
 apply (Some (ty11 :->: ty12) t1) (Some ty2 t2)
   | Just Refl <- testEquality ty11 ty2 = return $ Some ty12 (t1 $$ t2)
+-- (World T -> ...) applied to T: promote the argument to (World T) with const
 apply (Some (TTyWorld ty11 :->: ty12) t1) (Some ty2 t2)
   | Just Refl <- testEquality ty11 ty2 = return $ Some ty12 (t1 $$ (K .$$ t2))
+-- (S -> T) applied to (World S): lift the function to (World S -> World T).
 apply (Some (ty11 :->: ty12) t1) (Some (TTyWorld ty2) t2)
   | Just Refl <- testEquality ty11 ty2 = return $ Some (TTyWorld ty12) (B .$$ t1 $$ t2)
+-- World (S -> T) applied to S.  Note this case and the next are
+-- needed because in the previous case, when (S -> T) is lifted to
+-- (World S -> World T), T may itself be a function type.
 apply (Some (TTyWorld (ty11 :->: ty12)) t1) (Some ty2 t2)
   | Just Refl <- testEquality ty11 ty2 = return $ Some (TTyWorld ty12) (S .$$ t1 $$ (K .$$ t2))
+-- World (S -> T) applied to (World S)
 apply (Some (TTyWorld (ty11 :->: ty12)) t1) (Some (TTyWorld ty2) t2)
   | Just Refl <- testEquality ty11 ty2 = return $ Some (TTyWorld ty12) (S .$$ t1 $$ t2)
 apply t1 t2 = throwError $ ApplyErr t1 t2
@@ -309,6 +391,17 @@ apply t1 t2 = throwError $ ApplyErr t1 t2
 applyTo :: (Has (Throw CheckErr) sig m) => Some (TTerm g) -> Some (TTerm g) -> m (Some (TTerm g))
 applyTo = flip apply
 
+-- | Infer the type of an operator: turn a raw operator into a
+--   type-indexed constant.  However, some operators are polymorphic,
+--   so we also provide a list of type arguments.  For example, the
+--   type of the negation operator can be either (Int -> Int) or
+--   (Float -> Float) so we provide it as an argument.
+--
+--   Currently, all operators take at most one type argument, so
+--   (Maybe SomeTy) might seem more appropriate than [SomeTy], but
+--   that is just a coincidence; in general one can easily imagine
+--   operators that are polymorphic in more than one type variable,
+--   and we may wish to add such in the future.
 inferOp :: (Has (Throw CheckErr) sig m) => [SomeTy] -> Op -> m (Some (TTerm g))
 inferOp _ Not = return $ Some (TTyBool :->: TTyBool) (embed CNot)
 inferOp [SomeTy tyA] Neg = Some (tyA :->: tyA) <$> checkNum tyA (return $ embed CNeg)
@@ -337,6 +430,18 @@ inferOp [SomeTy tyA] Mask = Some (TTyWorld TTyBool :->: TTyWorld tyA :->: TTyWor
 inferOp [SomeTy tyA] Overlay = Some (tyA :->: tyA :->: tyA) <$> checkOver tyA (return $ embed COver)
 inferOp tys op = error $ "bad call to inferOp: " ++ show tys ++ " " ++ show op
 
+-- | Given a raw operator and the terms the operator is applied to,
+--   select which types should be supplied as the type arguments to
+--   the operator.  For example, for an operator like @+@ we can just
+--   select the type of its first argument; for an operator like @if@,
+--   we must select the type of its second argument, since @if : Bool
+--   -> a -> a -> a@.  In all cases we must also select the underlying
+--   base type in case the argument has a @World@ type.  For example
+--   if @+@ is applied to an argument of type @World Int@ we still
+--   want to give @+@ the type @Int -> Int -> Int@.  It can be lifted
+--   to have type @World Int -> World Int -> World Int@ but that will
+--   be taken care of by application, which will insert the right
+--   combinators to do the lifting.
 typeArgsFor :: Op -> [Some (TTerm g)] -> [SomeTy]
 typeArgsFor op (t : _)
   | op `elem` [Neg, Abs, Add, Sub, Mul, Div, Mod, Eq, Neq, Lt, Leq, Gt, Geq] = [getBaseType t]
@@ -346,20 +451,22 @@ typeArgsFor op (_ : t : _)
   | op `elem` [If, Mask, Overlay] = [getBaseType t]
 typeArgsFor _ _ = []
 
+-- | Typecheck the application of an operator to some terms, returning
+--   a typed, elaborated version of the application.
 applyOp ::
   ( Has (Throw CheckErr) sig m
   , Has (Reader EntityMap) sig m
   , Has (Reader WExpMap) sig m
   ) =>
   Ctx g ->
-  ([Some (TTerm g)] -> [SomeTy]) ->
   Op ->
   [WExp] ->
   m (Some (TTerm g))
-applyOp ctx typeArgs op ts = do
+applyOp ctx op ts = do
   tts <- mapM (infer ctx) ts
-  foldl (\r -> (r >>=) . applyTo) (inferOp (typeArgs tts) op) tts
+  foldl (\r -> (r >>=) . applyTo) (inferOp (typeArgsFor op tts) op) tts
 
+-- | Infer the type of a term, and elaborate along the way.
 infer ::
   forall sig m g.
   ( Has (Throw CheckErr) sig m
@@ -376,20 +483,21 @@ infer _ (WCell c) = do
   c' <- resolveCell c
   return $ Some TTyCell (embed (CCell c'))
 infer ctx (WVar x) = mapSome TVar <$> lookup x ctx
-infer ctx (WOp op ts) = applyOp ctx (typeArgsFor op) op ts
+infer ctx (WOp op ts) = applyOp ctx op ts
 infer _ WSeed = return $ Some TTyInt (embed CSeed)
 infer _ (WCoord ax) = return $ Some (TTyWorld TTyInt) (embed (CCoord ax))
 infer _ WHash = return $ Some (TTyWorld TTyInt) (embed CHash)
 infer ctx (WLet defs body) = inferLet ctx defs body
 infer ctx (WOverlay ts) = inferOverlay ctx ts
-infer _ctx (WCat _ax _ts) = undefined
-infer _ctx (WStruct _pal _rect) = undefined
 infer _ctx (WImport key) = do
   wexpMap <- ask @WExpMap
   case M.lookup key wexpMap of
     Just (Some ty t) -> return (Some ty (weaken @g t))
     Nothing -> throwError (UnknownErr 14)
 
+-- | Try to resolve a 'RawCellVal'---containing only 'Text' names for
+--   terrain, entities, and robots---into a real 'CellVal' with
+--   references to actual terrain, entities, and robots.
 resolveCell ::
   (Has (Throw CheckErr) sig m, Has (Reader EntityMap) sig m) =>
   RawCellVal ->
@@ -398,6 +506,8 @@ resolveCell items = do
   cellVals <- mapM resolveCellItem items
   return $ foldl' (<+>) empty cellVals
 
+-- | Try to resolve one cell item name into an actual item (terrain,
+-- entity, robot, etc.).
 resolveCellItem ::
   forall sig m.
   (Has (Throw CheckErr) sig m, Has (Reader EntityMap) sig m) =>
@@ -429,6 +539,8 @@ resolveCellItem (mCellTag, item) = case mCellTag of
           return . fmap mkEntity $ lookupEntityName eName em
     CellRobot -> \_ -> return Nothing -- XXX robots!
 
+-- | Infer the type of a let expression, and elaborate into a series
+--   of lambda applications.
 inferLet ::
   ( Has (Throw CheckErr) sig m
   , Has (Reader EntityMap) sig m
@@ -444,6 +556,8 @@ inferLet ctx ((x, e) : xs) body = do
   Some ty2 let' <- inferLet (CCons x ty1 ctx) xs body
   apply (Some (ty1 :->: ty2) (TLam let')) e'
 
+-- | Infer the type of an @overlay@ expression, and elaborate into a
+--   chain of @<>@ (over) operations.
 inferOverlay ::
   ( Has (Throw CheckErr) sig m
   , Has (Reader EntityMap) sig m
@@ -453,7 +567,9 @@ inferOverlay ::
   NE.NonEmpty WExp ->
   m (Some (TTerm g))
 inferOverlay ctx es = case NE.uncons es of
+  -- @overlay [e] = e@
   (e, Nothing) -> infer ctx e
+  -- @overlay (e : es') = e <> overlay es'@
   (e, Just es') -> do
     e' <- infer ctx e
     o' <- inferOverlay ctx es'
